Revert "Implement QTest::qWait() in terms of QTest::qWaitFor()"
This reverts commit 5c908c826313143a65ef7e95e5af6625fe1ba813.

git-bisect points to this commit as the cause for serious test
regressions in the qtwebengine module where test time execution
goes up 10 times (at least for debug builds), causing timeouts.

The reason for the time regression in test execution is caused
as the 'processEvents' call is no longer executed with
'remaining' time:

QCoreApplication::processEvents(QEventLoop::AllEvents, remaining)

'processEvents' do not spin for the whole duration and instead it calls
'predicate' after all event processing but before proceeding with new
events. This introduces significant Chromium's message pump lag
and makes test execution much slower.

In case of relanding this change we need to go through all tests and
extend timeouts, which is not feasible at the moment.

This is a quick-fix for 6.3.

Pick-to: dev
Change-Id: I90696479bfb9f0a0b8a8acc5bb7e7058b7d0c462
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>