Stabilize tst_QTouchEvent::multiPointRawEventTranslationOnTouchPad
The test is impacted by the position of the mouse cursor, and fails if
the mouse enters the test window when it gets shown. Try to move the
cursor away from the window.

As a drive-by, const'ify some of the local variables.

Revert 7b4b5115ddb2d492c476942090a3fa7992e9fa7d by removing the
blacklisting of the test on macOS in CI.

Change-Id: Ia0c554fdf161fd4eb4aa3965e937c7db8ceeef8f
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit dd3c4cb9bfc965ae40e914216862fbb350c55992)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>